fix(dmg): improve handling of AEA-encrypted DMGs and cleanup logic

+++ internal/commands/ent/ent.go
@@ -313,35 +313,50 @@ func DiffDatabases(db1, db2 map[string]string, conf *Config) (string, error) {
 }
 
 func scanEnts(ipswPath, dmgPath, dmgType string, conf *Config) (map[string]string, error) {
-	// check if filesystem DMG already exists (due to previous mount command)
-	if _, err := os.Stat(dmgPath); os.IsNotExist(err) {
-		dmgs, err := utils.Unzip(ipswPath, "", func(f *zip.File) bool {
-			return strings.EqualFold(filepath.Base(f.Name), dmgPath)
-		})
-		if err != nil {
-			return nil, fmt.Errorf("failed to extract %s from IPSW: %v", dmgPath, err)
+	skipCleanup := false
+
+	// For AEA-encrypted DMGs, check if the decrypted version already exists
+	// (e.g. already extracted + mounted by a prior step like mountSystemOsDMGs).
+	// Reuse it to avoid overwriting a mounted DMG's backing file.
+	if filepath.Ext(dmgPath) == ".aea" {
+		decryptedPath := strings.TrimSuffix(dmgPath, filepath.Ext(dmgPath))
+		if _, err := os.Stat(decryptedPath); err == nil {
+			dmgPath = decryptedPath
+			skipCleanup = true
 		}
-		if len(dmgs) == 0 {
-			return nil, fmt.Errorf("failed to find %s in IPSW", dmgPath)
-		}
-		defer os.Remove(dmgs[0])
-	} else {
-		utils.Indent(log.Debug, 2)(fmt.Sprintf("Found extracted %s", dmgPath))
 	}
 
-	if filepath.Ext(dmgPath) == ".aea" {
-		var err error
-		dmgPath, err = aea.Decrypt(&aea.DecryptConfig{
-			Input:    dmgPath,
-			Output:   filepath.Dir(dmgPath),
-			PemDB:    conf.PemDB,
-			Proxy:    "",    // TODO: make proxy configurable
-			Insecure: false, // TODO: make insecure configurable
-		})
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse AEA encrypted DMG: %v", err)
+	if !skipCleanup {
+		// check if filesystem DMG already exists (due to previous mount command)
+		if _, err := os.Stat(dmgPath); os.IsNotExist(err) {
+			dmgs, err := utils.Unzip(ipswPath, "", func(f *zip.File) bool {
+				return strings.EqualFold(filepath.Base(f.Name), dmgPath)
+			})
+			if err != nil {
+				return nil, fmt.Errorf("failed to extract %s from IPSW: %v", dmgPath, err)
+			}
+			if len(dmgs) == 0 {
+				return nil, fmt.Errorf("failed to find %s in IPSW", dmgPath)
+			}
+			defer os.Remove(dmgs[0])
+		} else {
+			utils.Indent(log.Debug, 2)(fmt.Sprintf("Found extracted %s", dmgPath))
 		}
-		defer os.Remove(dmgPath)
+
+		if filepath.Ext(dmgPath) == ".aea" {
+			var err error
+			dmgPath, err = aea.Decrypt(&aea.DecryptConfig{
+				Input:    dmgPath,
+				Output:   filepath.Dir(dmgPath),
+				PemDB:    conf.PemDB,
+				Proxy:    "",    // TODO: make proxy configurable
+				Insecure: false, // TODO: make insecure configurable
+			})
+			if err != nil {
+				return nil, fmt.Errorf("failed to parse AEA encrypted DMG: %v", err)
+			}
+			defer os.Remove(dmgPath)
+		}
 	}
 
 	utils.Indent(log.Debug, 2)(fmt.Sprintf("Mounting %s %s", dmgType, dmgPath))
